feat(release): model registration + schema drift tooling
A read-only recon run came first, as planned, and it changed the design. A strict
"compare_metadata() must be empty" test would have needed a ~395-entry allowlist,
which is not a check -- it is a changelog, and it would have become the next
expected-schemas.tsv. Categorising the diffs showed ~295 index and ~75 constraint
entries come from ONE cosmetic cause: hand-written raw-SQL migrations name
indexes idx_<table>_<col> while SQLAlchemy's implicit name for index=True is
ix_<table>_<col>. Both indexes exist and cover the same column, and
app/db/base.py has no naming convention on purpose (adding one renames every
existing constraint).

So the gate is scoped to the categories that actually raise at runtime, and the
rest is reported.

Always-on, zero-noise:
- test_model_registration.py asserts every module under app/models that declares
  a __tablename__ is reachable from Base.metadata after `import app.models`, and
  that the import line exists in __init__.py rather than happening transitively.

Real bugs this found, fixed here:
- app/models/system_settings.py was never imported by app/models/__init__.py, so
  `import app.models` registered 52 tables instead of 53. Base.metadata was
  incomplete, meaning autogenerate and create_all silently omitted the table --
  and app/models/CLAUDE.md already documents __init__.py as "the canonical import
  surface. A new model must be added here."
- MediaFile.file_size was declared Integer while the column is bigint and the app
  advertises 15 GB uploads (MAX_UPLOAD_BYTES), well past Integer's 2.1 GB ceiling.
  Reads worked (Python ints are arbitrary precision) so nothing failed, but the
  model was wrong and DDL generated from it would have been too narrow. The
  sibling models already use BigInteger.

Reporting:
- scripts/check-schema-drift.py groups diffs into tables / columns / types /
  indexes / constraints so the signal categories are readable instead of buried,
  with --json and an optional --fail-on. Its header documents why the raw count
  is not currently meaningful and what would make it so.
- test_schema_drift.py gates on tables and columns only, behind
  RUN_SCHEMA_DRIFT_TESTS (same convention as the other service-dependent suites).

It currently reports 4 genuine pre-existing issues, left for a follow-up because
each needs a migration rather than a code change: upload_session,
speaker_audio_clip and user_certificate_preferences exist in the database with no
model and no reference anywhere under app/ (leftovers from removed features), and
tag.created_at exists in the database but not in the model.

+++ backend/app/models/media.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from typing import TYPE_CHECKING
 from typing import Any
 
+from sqlalchemy import BigInteger
 from sqlalchemy import Boolean
 from sqlalchemy import DateTime
 from sqlalchemy import Enum as SAEnum
@@ -54,7 +55,14 @@ class MediaFile(Base):
         DateTime(timezone=True), nullable=True
     )  # When processing completed
     duration: Mapped[float | None] = mapped_column(Float, nullable=True)  # Duration in seconds
-    file_size: Mapped[int] = mapped_column(Integer, nullable=False)  # Size in bytes
+    # BigInteger, not Integer: the column is bigint in Postgres and the app
+    # advertises 15 GB uploads (MAX_UPLOAD_BYTES), well past Integer's 2.1 GB
+    # ceiling. The model said Integer while the database said bigint — reads
+    # still worked (Python ints are arbitrary precision), but the model was
+    # wrong, DDL generated from it would have been too narrow, and the sibling
+    # models (pipeline_timing.file_size_bytes, watch_source) already use
+    # BigInteger. Found by scripts/check-schema-drift.py.
+    file_size: Mapped[int] = mapped_column(BigInteger, nullable=False)  # Size in bytes
     content_type: Mapped[str] = mapped_column(String, nullable=False)  # MIME type
     is_public: Mapped[bool | None] = mapped_column(
         Boolean, default=False
